+++ src/main/java/sortBgzipIndex/MakeTabixIndex.java
@@ -16,6 +16,7 @@ import java.sql.Statement;
 
 import org.sqlite.SQLiteException;
 
+
 import exceptions.InvalidRecordException;
 import htsjdk.samtools.util.BlockCompressedOutputStream;
 import htsjdk.samtools.util.CloserUtil;
@@ -37,6 +38,8 @@ import utils.GtfLine;
 public class MakeTabixIndex {
 
     private File sqliteFile;
+    final private String columnSeparator;
+    final private TabixFormat tabixFormat;
     
     /** Sort, block compress and index the input with format fmt to the given output file.
      * Input is either a local file, possibly compressed, or a URL.
@@ -45,20 +48,26 @@ public class MakeTabixIndex {
      * @throws SQLException 
      * @throws ClassNotFoundException 
      * */
-    public MakeTabixIndex(String intab, File bgzfOut, TabixFormat fmt) throws IOException, InvalidRecordException, ClassNotFoundException, SQLException{
+    public MakeTabixIndex(String intab, File bgzfOut, TabixFormat tabixFormat) throws IOException, InvalidRecordException, ClassNotFoundException, SQLException{
+        
+        this.tabixFormat = tabixFormat;
+        this.columnSeparator = Utils.getColumnSeparator(intab);
         
         File tmp = Utils.createTempFile(".asciigenome", "makeTabixIndex.tmp.gz", true);
         File tmpTbi= new File(tmp.getAbsolutePath() + FileExtensions.TABIX_INDEX);
         tmpTbi.deleteOnExit();
-        
+
         try{
+            if(!this.columnSeparator.equals("\t")) {
+                throw new InvalidRecordException();
+            }
             // Try to block compress and create index assuming the file is sorted
-            blockCompressAndIndex(intab, tmp, fmt);
+            blockCompressAndIndex(intab, tmp);
         } catch(Exception e){
             // If intab is not sorted, sort it first.
             File sorted= Utils.createTempFile(".asciigenome.", ".sorted.tmp", true);
-            sortByChromThenPos(intab, sorted, fmt);
-            blockCompressAndIndex(sorted.getAbsolutePath(), tmp, fmt);
+            sortByChromThenPos(intab, sorted);
+            blockCompressAndIndex(sorted.getAbsolutePath(), tmp);
             Files.delete(Paths.get(sorted.getAbsolutePath()));
         }
         
@@ -82,20 +91,17 @@ public class MakeTabixIndex {
      * @throws IOException 
      * @throws InvalidRecordException 
      * */
-    private void blockCompressAndIndex(String intab, File bgzfOut, TabixFormat fmt) throws IOException, InvalidRecordException {
-                
+    private void blockCompressAndIndex(String intab, File bgzfOut) throws IOException, InvalidRecordException {
         BlockCompressedOutputStream writer = new BlockCompressedOutputStream(bgzfOut);
         long filePosition= writer.getFilePointer();
             
-        TabixIndexCreator indexCreator=new TabixIndexCreator(fmt);
-        
-        // boolean first= true;
+        TabixIndexCreator indexCreator=new TabixIndexCreator(this.tabixFormat);
         
         // This is relevant to vcf files only: Prepare header and codec
         // ------------------------------------------------------------
         VCFHeader vcfHeader= null;
         VCFCodec vcfCodec= null;
-        if(fmt.equals(TabixFormat.VCF)){
+        if(this.tabixFormat.equals(TabixFormat.VCF)){
             try{
                 VCFFileReader vcfr= new VCFFileReader(new File(intab), false);
                 vcfHeader= vcfr.getFileHeader();
@@ -110,30 +116,26 @@ public class MakeTabixIndex {
 
         int nWarnings= 10;
         LineIterator lin= utils.IOUtils.openURIForLineIterator(intab);
+        boolean dataLinesFound = false;
         while(lin.hasNext()){
             
             String line = lin.next().trim();
-            
-            try{
-                if(line.isEmpty() || line.startsWith("track ")){
-                    continue;
-                }
-                if(line.startsWith("#")){
-                    writer.write((line + "\n").getBytes());
-                    filePosition = writer.getFilePointer();
-                    continue;
-                }
-                if(line.startsWith("##FASTA")){
-                    break;
-                }			
-
-                addLineToIndex(line, indexCreator, filePosition, fmt, vcfHeader, vcfCodec);
-                
-                writer.write(line.getBytes());
-                writer.write('\n');
+            if(line.isEmpty() || line.startsWith("track ")){
+                continue;
+            }
+            if(line.startsWith("#")){
+                writer.write((line + "\n").getBytes());
                 filePosition = writer.getFilePointer();
+                continue;
+            }
+            if(line.startsWith("##FASTA") && dataLinesFound){
+                break;
+            }			
+            try {
+                addLineToIndex(line, indexCreator, filePosition, vcfHeader, vcfCodec);
             } catch(Exception e){
-                if(e.getMessage().contains("added out sequence of order") || e.getMessage().contains("Features added out of order")){
+                if(e.getMessage().contains("added out sequence of order") || 
+                        e.getMessage().contains("Features added out of order")){
                     // Get a string marker for out-of-order from htsjdk/tribble/index/tabix/TabixIndexCreator.java 
                     writer.close();
                     throw new InvalidRecordException();
@@ -146,6 +148,10 @@ public class MakeTabixIndex {
                 }
                 nWarnings--;
             }
+            writer.write(line.getBytes());
+            writer.write('\n');
+            filePosition = writer.getFilePointer();
+            dataLinesFound = true;
         }
 
         writer.flush();
@@ -158,25 +164,19 @@ public class MakeTabixIndex {
 
     /** Set vcfHeader and vcfCodec to null if reading non-vcf line.
      * */
-    private void addLineToIndex(String line, TabixIndexCreator indexCreator, long filePosition, TabixFormat fmt, VCFHeader vcfHeader, VCFCodec vcfCodec) throws InvalidRecordException {
-
-        if(fmt.equals(TabixFormat.BED)){
-            
+    private void addLineToIndex(String line, TabixIndexCreator indexCreator, long filePosition, VCFHeader vcfHeader, VCFCodec vcfCodec) throws InvalidRecordException {
+        if(this.tabixFormat.equals(TabixFormat.BED)){
             BedLineCodec bedCodec= new BedLineCodec();
             BedLine bed = bedCodec.decode(line);
             indexCreator.addFeature(bed, filePosition);
-        
-        } else if(fmt.equals(TabixFormat.GFF)){
+        } else if(this.tabixFormat.equals(TabixFormat.GFF)){
             GtfLine gtf= new GtfLine(line.split("\t"));
             indexCreator.addFeature(gtf, filePosition);
-        
-        } else if(fmt.equals(TabixFormat.VCF)) {
-
+        } else if(this.tabixFormat.equals(TabixFormat.VCF)) {
             VariantContext vcf = vcfCodec.decode(line);
             indexCreator.addFeature(vcf, filePosition);
-            
         } else {
-            System.err.println("Unexpected TabixFormat: " + fmt.sequenceColumn + " " + fmt.startPositionColumn);
+            System.err.println("Unexpected TabixFormat: " + this.tabixFormat.sequenceColumn + " " + this.tabixFormat.startPositionColumn);
             throw new InvalidRecordException();
         }	
     }
@@ -185,15 +185,15 @@ public class MakeTabixIndex {
      * indexes for the chrom and pos column. For bed use 1 and 2 respectively. For use GTF/GFF  1 and 4.
      * Comment lines, starting with #, are returned as they are. Reading stops if the line ##FASTA is found.
      * */
-    private void sortByChromThenPos(String unsorted, File sorted, TabixFormat fmt) throws SQLException, InvalidRecordException, IOException, ClassNotFoundException{
+    private void sortByChromThenPos(String unsorted, File sorted) throws SQLException, InvalidRecordException, IOException, ClassNotFoundException{
 
         int chromIdx= 1;
         int posIdx= 2;
-        if(fmt.equals(TabixFormat.BED)){
+        if(this.tabixFormat.equals(TabixFormat.BED)){
             //
-        } else if(fmt.equals(TabixFormat.GFF)){
+        } else if(this.tabixFormat.equals(TabixFormat.GFF)){
             posIdx= 4;
-        } else if(fmt.equals(TabixFormat.VCF)){
+        } else if(this.tabixFormat.equals(TabixFormat.VCF)){
             posIdx= 2;
         } else {
             System.err.println("Invalid format found");
@@ -214,24 +214,43 @@ public class MakeTabixIndex {
         BufferedReader br= Utils.reader(unsorted);
         BufferedWriter wr= new BufferedWriter(new FileWriter(sorted));
         String line;
+        int n = 0;
+        boolean dataLinesfound = false;
         while((line = br.readLine()) != null){
-            if(line.trim().startsWith("##FASTA")){
+            if(line.trim().startsWith("##FASTA") && dataLinesfound && this.tabixFormat.equals(TabixFormat.GFF)){
                 break;
             }
             if(line.trim().startsWith("#")){
                 wr.write(line + "\n");
                 continue;
             }
-            String[] tabs= line.split("\t"); 
+            if(line.trim().isEmpty()){
+                continue;
+            }
+            if(this.tabixFormat.equals(TabixFormat.BED) && !this.columnSeparator.equals("\t")) {
+                line = line.replace(this.columnSeparator, "\t");
+            }
+            String[] tabs = line.split("\t");
+            if(n == 0 && this.tabixFormat.equals(TabixFormat.BED)) {
+                // Allow first uncommented line to fail
+                n++;
+                try {
+                    Integer.parseInt(tabs[1]);
+                    Integer.parseInt(tabs[2]);
+                } catch(NumberFormatException e) {
+                    continue;
+                }
+            }
             stmtInsert.setString(1, tabs[chromIdx-1]);
             stmtInsert.setInt(2, Integer.parseInt(tabs[posIdx-1]));
-            if(fmt.equals(TabixFormat.VCF)){
+            if(this.tabixFormat.equals(TabixFormat.VCF)){
                 stmtInsert.setInt(3, 0);
             } else {
                 stmtInsert.setInt(3, Integer.parseInt(tabs[posIdx]));
             }
             stmtInsert.setString(4, line.replaceAll("\n$", ""));
             stmtInsert.executeUpdate();
+            dataLinesfound = true;
         }
         stmtInsert.close();
         br.close();